% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combination_count.R
\name{combination_count}
\alias{combination_count}
\title{combination_count}
\usage{
combination_count(N, n)
}
\arguments{
\item{N}{Input value representing the total number of elements.}

\item{n}{Input value representing the number of elements to be selected.}
}
\value{
the number of possible combinations when selecting n elements from N
}
\description{
Gets the number of combination when selecting n elements from N
}
\examples{
combination_count(5,2)

}
